fuel_route/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-decouple==3.8
requests==2.32.5
gunicorn==21.2.0
django-redis==5.4.0
redis==5.2.1
